Remove `AttributeLintKind::DocUnknownInclude` variant
